Chameleon: the OS is now also a GPT-partitioned disk image (with a small quirk)
The quirk is that the ZIP trailer is inserted after the end of the
GPT-partitioned space, but GPT normally covers the whole disk.

This image is intended to be burned on a CD, flashed onto a floppy, USB key or
SD card. In these last two cases (where the GPT would make sense), the end of
the disk cannot be known in advance, so an incorrectly-placed GPT mirror is
not quite unexpected.